requirements.txt
tensorflow
faiss-cpu
numpy
scikit-learn